build: Improve fuzz tests' reliability Establish conventions which aligns with what is supported upstream today, whilst expanding on documentation to ensure other contributors have pointers on how to debug/check for issues going forwards.
