+++ runtime/conditions/patch_test.go
@@ -27,7 +27,6 @@ import (
 	"testing"
 	"time"
 
-	fuzz "github.com/AdaLogics/go-fuzz-headers"
 	. "github.com/onsi/gomega"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -289,50 +288,4 @@ func TestApplyDoesNotAlterLastTransitionTime(t *testing.T) {
 
 	g.Expect(err).ToNot(HaveOccurred())
 	g.Expect(latest.GetConditions()).To(Equal(after.GetConditions()))
-}
-
-func Fuzz_PatchApply(f *testing.F) {
-	f.Fuzz(func(t *testing.T,
-		beforeData, afterData, setterData []byte) {
-
-		before, err := newFake(fuzz.NewConsumer(beforeData))
-		if err != nil {
-			return
-		}
-
-		after, err := newFake(fuzz.NewConsumer(afterData))
-		if err != nil {
-			return
-		}
-
-		patch := NewPatch(before, after)
-		setter, err := newFake(fuzz.NewConsumer(setterData))
-		if err != nil {
-			return
-		}
-
-		_ = patch.Apply(setter)
-	})
-}
-
-func newFake(fc *fuzz.ConsumeFuzzer) (*testdata.Fake, error) {
-	obj := &testdata.Fake{}
-	noOfConditions, err := fc.GetInt()
-	if err != nil {
-		return obj, err
-	}
-
-	maxNoOfConditions := 30
-	conditions := make([]metav1.Condition, 0)
-	for i := 0; i < noOfConditions%maxNoOfConditions; i++ {
-		c := metav1.Condition{}
-		err = fc.GenerateStruct(&c)
-		if err != nil {
-			return obj, err
-		}
-
-		conditions = append(conditions, c)
-	}
-	obj.SetConditions(conditions)
-	return obj, nil
 }
